Show loader while fetching posts
