+++ src/App.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import Container from 'react-bootstrap/Container';
 import './App.scss';
-import PostsList from './features/posts/PostList';
+import Posts from './features/posts/Posts';
 
 function App() {
   return (
     <Container>
-      <PostsList />
+      <Posts />
     </Container>
   );
 }
