+++ src/features/posts/PostTable.module.scss
